+++ checks/dns_root_servers.go
@@ -14,31 +14,16 @@ type DNSRootServersCheck struct {
 	netiscopeCheckBase
 }
 
-type serverAddress struct {
-	v4 string
-	v6 string
-}
-type rootDNSServer struct {
-	letter    string
-	addresses serverAddress
+type rootDNSServerCheckType struct {
+	Letter    string
+	IPv4      string
+	IPv6      string
+	Pingable4 bool
+	Pingable6 bool
 }
 
 // these really don't change too often
-var rootDNSServers = []rootDNSServer{
-	{"A", serverAddress{"198.41.0.4", "2001:503:ba3e::2:30"}},   // Verisign, Inc.
-	{"B", serverAddress{"170.247.170.2", "2801:1b8:10::b"}},     // Information Sciences Institute (ISI)
-	{"C", serverAddress{"192.33.4.12", "2001:500:2::c"}},        // Cogent Communications
-	{"D", serverAddress{"199.7.91.13", "2001:500:2d::d"}},       // University of Maryland
-	{"E", serverAddress{"192.203.230.10", "2001:500:a8::e"}},    // NASA Ames Research Center
-	{"F", serverAddress{"192.5.5.241", "2001:500:2f::f"}},       // Internet Systems Consortium, Inc. (ISC)
-	{"G", serverAddress{"192.112.36.4", "2001:500:12::d0d"}},    // Defense Information Systems Agency
-	{"H", serverAddress{"198.97.190.53", "2001:500:1::53"}},     // U.S. Army Research Lab
-	{"I", serverAddress{"192.36.148.17", "2001:7fe::53"}},       // Netnod
-	{"J", serverAddress{"192.58.128.30", "2001:503:c27::2:30"}}, // Verisign, Inc.
-	{"K", serverAddress{"193.0.14.129", "2001:7fd::1"}},         // RIPE NCC
-	{"L", serverAddress{"199.7.83.42", "2001:500:9f::42"}},      // ICANN
-	{"M", serverAddress{"202.12.27.33", "2001:dc3::35"}},        // WIDE Project
-}
+var rootDNSServers []rootDNSServerCheckType
 
 // Start executes the DNS root server check
 func (check *DNSRootServersCheck) start() {
@@ -49,16 +34,39 @@ func (check *DNSRootServersCheck) start() {
 			break
 		}
 		if !util.SkipIPv4() {
-			checkRootDNSServer(&check.netiscopeCheckBase, server.letter, "IPv4", server.addresses.v4)
+			checkRootDNSServer(&check.netiscopeCheckBase, server.Letter, "IPv4", server.IPv4, server.Pingable4)
 		}
 		if !util.SkipIPv6() {
-			checkRootDNSServer(&check.netiscopeCheckBase, server.letter, "IPv6", server.addresses.v6)
+			checkRootDNSServer(&check.netiscopeCheckBase, server.Letter, "IPv6", server.IPv6, server.Pingable6)
 		}
 	}
 
 	// TODO: compare responses from different root servers?
 
 	check.netiscopeCheckBase.finish()
+}
+
+func (check *DNSRootServersCheck) configure() {
+	for _, item := range util.LoadRootDNSServerData() {
+		if len(item) < 3 {
+			check.log(
+				LogLevelError,
+				"DNS_ROOT_SERVER_CONFIG_ERROR",
+				fmt.Sprintf(
+					"Wrong configuration item for DNS root server check: %v",
+					item,
+				),
+			)
+		}
+		root := rootDNSServerCheckType{
+			Letter:    item[0],
+			IPv4:      item[1],
+			IPv6:      item[2],
+			Pingable4: len(item) == 5 && item[3] == "true",
+			Pingable6: len(item) == 5 && item[4] == "true",
+		}
+		rootDNSServers = append(rootDNSServers, root)
+	}
 }
 
 // check a particular DNS root server on IPv4 or IPv6
@@ -70,6 +78,7 @@ func checkRootDNSServer(
 	letter string,
 	af string,
 	server string,
+	pingable bool,
 ) {
 	check.log(
 		LogLevelInfo,
@@ -79,7 +88,7 @@ func checkRootDNSServer(
 			letter, af, server,
 		),
 	)
-	testRootDNSServerOnAddressFamily(check, letter, af, server)
+	testRootDNSServerOnAddressFamily(check, letter, af, server, pingable)
 }
 
 // test a root DNS server on a particular address family
@@ -91,8 +100,9 @@ func testRootDNSServerOnAddressFamily(
 	letter string,
 	af string,
 	server string,
+	pingable bool,
 ) {
-	if shouldCheckDNSFunction("ping") {
+	if pingable && shouldCheckDNSFunction("ping") {
 		reportResolversOnAddressFamily(
 			check,
 			"ROOT_DNS_SERVER", af, letter+"-root DNS server", "PING", "reachable", []string{server},